corregir

+++ estudiantes/src/main/java/aed/Estudiante.java
@@ -7,7 +7,7 @@ public class Estudiante {
     int fila;
     int columna;
     int correctas;
-    int nota;
+    double nota;
     boolean entrego;
     boolean sospechoso;
     Examen examen;
@@ -19,7 +19,7 @@ public class Estudiante {
     public Estudiante(int id, int cantPreguntas, int ladoAula){
         this.id = id;                                                           // O(1)                        // O(1)
         this.correctas = 0;                                                     // O(1)
-        this.nota = 0;                                                          // O(1)
+        this.nota = 0.0;                                                          // O(1)
         this.entrego = false;                                                   // O(1)
         this.sospechoso = true;                                                 // O(1)
         this.examen = new Examen(cantPreguntas);                                // O(R)
